Added all particle visualisation

+++ single_sphere_all_types.py
@@ -146,7 +146,9 @@ def single_sphere(reg, snap, part_type, soft, t=0, p=0, num=0):
     # rgb_output = blend2.Overlay()
     rgb_output = dmgas_output
 
-    os.remove('plots/spheres/all_parts_single_sphere_reg' + reg + '_snap' + snap + '_angle%05d.png'%num)
+    if 'plots/spheres/all_parts_single_sphere_reg' + reg + '_snap' + snap \
+            + '_angle%05d.png'%num in list(os.listdir(os.cwd)):
+        os.remove('plots/spheres/all_parts_single_sphere_reg' + reg + '_snap' + snap + '_angle%05d.png'%num)
 
     plt.imsave('plots/spheres/all_parts_single_sphere_reg' + reg + '_snap' + snap + '_angle%05d.png'%num, rgb_output)
